requirements.txt
alembic==1.4.2
bcrypt==3.2.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.14.2
chardet==3.0.4
Click==7.0
DateTime==4.3
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SeaSurf==0.2.2
Flask-SQLAlchemy==2.4.4
flask-talisman==0.7.0
gmplot==1.4.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.18.0
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
Mako==1.1.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
oauth2client==4.1.3
openpyxl==2.6.3
pdfrw==0.4
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.20
PyPDF2==1.26.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2019.2
requests==2.24.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.18
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.6
xlwt==1.3.0
yagmail==0.11.220
zope.interface==5.1.0
